# src/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Express Starter API",
    "description": "Production-ready Express.js REST API with TypeScript, MongoDB, Redis, and JWT authentication",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "badlamoussaab@gmail.com"
    },
    "license": {
      "name": "ISC"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/"
    }
  ],
  "tags": [
    {
      "name": "Authentication",
      "description": "User authentication and authorization endpoints"
    },
    {
      "name": "Health",
      "description": "Health check and monitoring endpoints"
    },
    {
      "name": "Files",
      "description": "File upload and management endpoints"
    }
  ],
  "paths": {
    "/api-docs/": {
      "get": {
        "description": "",
        "responses": {
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api-docs/json": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/health/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "503": {
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/health/ready": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "503": {
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/health/live": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "example": "any"
                  },
                  "password": {
                    "example": "any"
                  },
                  "stay": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "example": "any"
                  },
                  "password": {
                    "example": "any"
                  },
                  "firstName": {
                    "example": "any"
                  },
                  "lastName": {
                    "example": "any"
                  },
                  "stay": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": {
                    "example": "any"
                  },
                  "stay": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/": {
      "get": {
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "507f1f77bcf86cd799439011"
          },
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "email": {
            "type": "string",
            "example": "john.doe@example.com"
          },
          "role": {
            "type": "string",
            "example": "user"
          },
          "enable": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "type": "string",
            "example": "2025-01-15T10:30:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2025-01-15T10:30:00.000Z"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "john.doe@example.com"
          },
          "password": {
            "type": "string",
            "example": "Test@1234"
          },
          "stay": {
            "type": "boolean",
            "example": false
          }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "john.doe@example.com"
          },
          "password": {
            "type": "string",
            "example": "Test@1234"
          },
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "stay": {
            "type": "boolean",
            "example": false
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "LOGIN_SUCCESS"
          },
          "message": {
            "type": "string",
            "example": "Logged in successfully"
          },
          "data": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "RefreshRequest": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "stay": {
            "type": "boolean",
            "example": false
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "error"
          },
          "message": {
            "type": "string",
            "example": "ERROR_TYPE"
          },
          "error": {
            "type": "string",
            "example": "Detailed error message"
          }
        }
      },
      "HealthResponse": {
        "type": "object",
        "properties": {
          "uptime": {
            "type": "number",
            "example": 12345.67
          },
          "timestamp": {
            "type": "number",
            "example": 1234567890000
          },
          "status": {
            "type": "string",
            "example": "OK"
          },
          "checks": {
            "type": "object",
            "properties": {
              "mongodb": {
                "type": "string",
                "example": "UP"
              },
              "redis": {
                "type": "string",
                "example": "UP"
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Enter your bearer token in the format: Bearer <token>"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token",
        "description": "JWT token stored in cookie"
      }
    }
  }
}